started working on video tab
